+++ src/renderer/src/components/right-sidebar/CreateHostedReviewComposer.tsx
@@ -28,7 +28,7 @@ import {
 } from '../../../../shared/hosted-review'
 import { stripBaseRef } from './create-pull-request-base-ref-normalization'
 import type { HostedReviewStackParent } from './useHostedReviewStackParent'
-import type { DropdownActionKind, DropdownEntry } from './source-control-dropdown-items'
+import type { DropdownActionKind, DropdownEntry } from './source-control-dropdown-item-types'
 import { CreateHostedReviewComposerFields } from './CreateHostedReviewComposerFields'
 import { getCreateButtonLabel } from './create-hosted-review-button-label'
 import {
