+++ src/renderer/src/components/right-sidebar/source-control/commit/commit-area-types.ts
@@ -1,4 +1,4 @@
-import type { DropdownActionKind, DropdownEntry } from '../../source-control-dropdown-items'
+import type { DropdownActionKind, DropdownEntry } from '../../source-control-dropdown-item-types'
 import type { PrimaryAction, RemoteOpKind } from '../../source-control-primary-action'
 import type { SourceControlPushRecovery } from '../sync/push-recovery'
 import type {
